% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_functions.R
\docType{package}
\name{pgee.mixed}
\alias{pgee.mixed}
\alias{pgee.mixed-package}
\title{pgee.mixed: Penalized Generalized Estimating Equations for Bivariate Mixed Outcomes}
\description{
Perform simultaneous estimation and variable selection for correlated bivariate
mixed outcomes (one continuous outcome and one binary outcome per cluster) using
penalized generalized estimating equations. In addition, clustered Gaussian and binary
outcomes can also be modeled. The SCAD, MCP, and LASSO penalties are supported.
Cross-validation can be performed to find the optimal regularization parameter(s).
}
\references{
Deshpande, V., Dey, D. K., and Schifano, E. D. (2016). Variable
selection for correlated bivariate mixed outcomes using penalized generalized
estimating equations. Technical Report 16-23, Department of Statistics,
University of Connecticut, Storrs, CT.

Wang, L., Zhou, J., and Qu, A. (2012). Penalized generalized
  estimating equations for high-dimensional longitudinal data analysis.
  Biometrics, \strong{68}, 353–360.
}

